refactor: clean dashboard_layout
